temp - Remove name field from  NodeDraftInput
